# Adjust WASM execution costing