--- a/radix-engine-tests/assets/metering/cuttlefish/cost_crypto_utils.csv
+++ b/radix-engine-tests/assets/metering/cuttlefish/cost_crypto_utils.csv
@@ -1,18 +1,18 @@
-Total Cost (XRD)                                                           ,            1.04966855779,    100.0%
-- Execution Cost (XRD)                                                     ,               0.67196615,     64.0%
+Total Cost (XRD)                                                           ,            1.04592820779,    100.0%
+- Execution Cost (XRD)                                                     ,                0.6682258,     63.9%
 - Finalization Cost (XRD)                                                  ,                0.0102517,      1.0%
-- Storage Cost (XRD)                                                       ,            0.36745070779,     35.0%
+- Storage Cost (XRD)                                                       ,            0.36745070779,     35.1%
 - Tipping Cost (XRD)                                                       ,                        0,      0.0%
 - Royalty Cost (XRD)                                                       ,                        0,      0.0%
-Execution Cost Breakdown                                                   ,                 13439323,    100.0%
+Execution Cost Breakdown                                                   ,                 13364516,    100.0%
 - AfterInvoke                                                              ,                      518,      0.0%
 - AllocateNodeId                                                           ,                     1455,      0.0%
 - BeforeInvoke                                                             ,                     7800,      0.1%
 - Blake2b256Hash                                                           ,                       91,      0.0%
 - Bls12381G2SignatureAggregate                                             ,                   243986,      1.8%
-- Bls12381V1AggregateVerify                                                ,                  2001824,     14.9%
+- Bls12381V1AggregateVerify                                                ,                  2001824,     15.0%
 - Bls12381V1FastAggregateVerify                                            ,                   632277,      4.7%
-- Bls12381V1Verify                                                         ,                   461378,      3.4%
+- Bls12381V1Verify                                                         ,                   461378,      3.5%
 - CheckReference                                                           ,                    80024,      0.6%
 - CloseSubstate                                                            ,                    18576,      0.1%
 - CreateNode                                                               ,                    13376,      0.1%
@@ -25,27 +25,27 @@ Execution Cost Breakdown                                                   ,    
 - MarkSubstateAsTransient                                                  ,                       55,      0.0%
 - OpenSubstate::GlobalFungibleResourceManager                              ,                   121872,      0.9%
 - OpenSubstate::GlobalGenericComponent                                     ,                    43690,      0.3%
-- OpenSubstate::GlobalPackage                                              ,                  3484661,     25.9%
+- OpenSubstate::GlobalPackage                                              ,                  3484661,     26.1%
 - OpenSubstate::InternalFungibleVault                                      ,                    90202,      0.7%
 - OpenSubstate::InternalGenericComponent                                   ,                    18682,      0.1%
 - OpenSubstate::InternalKeyValueStore                                      ,                    40536,      0.3%
 - PinNode                                                                  ,                      180,      0.0%
-- PrepareWasmCode                                                          ,                  2269642,     16.9%
+- PrepareWasmCode                                                          ,                  2269642,     17.0%
 - QueryActor                                                               ,                     1000,      0.0%
-- ReadSubstate                                                             ,                  2340609,     17.4%
+- ReadSubstate                                                             ,                  2340609,     17.5%
 - RunNativeCode::Worktop_drop                                              ,                    17918,      0.1%
 - RunNativeCode::get_amount_FungibleVault                                  ,                    14451,      0.1%
 - RunNativeCode::lock_fee                                                  ,                    45243,      0.3%
-- RunWasmCode::CryptoScrypto_blake2b_256_hash                              ,                    10893,      0.1%
-- RunWasmCode::CryptoScrypto_bls12381_g2_signature_aggregate               ,                   422302,      3.1%
-- RunWasmCode::CryptoScrypto_bls12381_v1_aggregate_verify                  ,                   296509,      2.2%
-- RunWasmCode::CryptoScrypto_bls12381_v1_fast_aggregate_verify             ,                   246475,      1.8%
-- RunWasmCode::CryptoScrypto_bls12381_v1_verify                            ,                    71371,      0.5%
-- RunWasmCode::CryptoScrypto_ed25519_verify                                ,                    50533,      0.4%
-- RunWasmCode::CryptoScrypto_keccak256_hash                                ,                    10884,      0.1%
-- RunWasmCode::CryptoScrypto_secp256k1_ecdsa_verify                        ,                    63316,      0.5%
-- RunWasmCode::CryptoScrypto_secp256k1_ecdsa_verify_and_key_recover        ,                    51265,      0.4%
-- RunWasmCode::Faucet_lock_fee                                             ,                    27840,      0.2%
+- RunWasmCode::CryptoScrypto_blake2b_256_hash                              ,                     9086,      0.1%
+- RunWasmCode::CryptoScrypto_bls12381_g2_signature_aggregate               ,                   400427,      3.0%
+- RunWasmCode::CryptoScrypto_bls12381_v1_aggregate_verify                  ,                   281194,      2.1%
+- RunWasmCode::CryptoScrypto_bls12381_v1_fast_aggregate_verify             ,                   233124,      1.7%
+- RunWasmCode::CryptoScrypto_bls12381_v1_verify                            ,                    66503,      0.5%
+- RunWasmCode::CryptoScrypto_ed25519_verify                                ,                    46414,      0.3%
+- RunWasmCode::CryptoScrypto_keccak256_hash                                ,                     8880,      0.1%
+- RunWasmCode::CryptoScrypto_secp256k1_ecdsa_verify                        ,                    58421,      0.4%
+- RunWasmCode::CryptoScrypto_secp256k1_ecdsa_verify_and_key_recover        ,                    47242,      0.4%
+- RunWasmCode::Faucet_lock_fee                                             ,                    25290,      0.2%
 - Secp256k1EcdsaVerify                                                     ,                    14705,      0.1%
 - Secp256k1EcdsaVerifyAndKeyRecover                                        ,                    14705,      0.1%
 - SetCallFrameData                                                         ,                      606,      0.0%