--- a/radix-engine-tests/assets/metering/cuttlefish/cost_flash_loan.csv
+++ b/radix-engine-tests/assets/metering/cuttlefish/cost_flash_loan.csv
@@ -1,10 +1,10 @@
-Total Cost (XRD)                                                           ,            0.62384462592,    100.0%
-- Execution Cost (XRD)                                                     ,                0.4634134,     74.3%
+Total Cost (XRD)                                                           ,            0.62335762592,    100.0%
+- Execution Cost (XRD)                                                     ,                0.4629264,     74.3%
 - Finalization Cost (XRD)                                                  ,                0.0322574,      5.2%
-- Storage Cost (XRD)                                                       ,            0.12817382592,     20.5%
+- Storage Cost (XRD)                                                       ,            0.12817382592,     20.6%
 - Tipping Cost (XRD)                                                       ,                        0,      0.0%
 - Royalty Cost (XRD)                                                       ,                        0,      0.0%
-Execution Cost Breakdown                                                   ,                  9268268,    100.0%
+Execution Cost Breakdown                                                   ,                  9258528,    100.0%
 - AfterInvoke                                                              ,                     1884,      0.0%
 - AllocateNodeId                                                           ,                     5141,      0.1%
 - BeforeInvoke                                                             ,                     7694,      0.1%
@@ -25,7 +25,7 @@ Execution Cost Breakdown                                                   ,    
 - OpenSubstate::InternalFungibleVault                                      ,                   290575,      3.1%
 - OpenSubstate::InternalGenericComponent                                   ,                   212121,      2.3%
 - PinNode                                                                  ,                      636,      0.0%
-- PrepareWasmCode                                                          ,                   967836,     10.4%
+- PrepareWasmCode                                                          ,                   967836,     10.5%
 - QueryActor                                                               ,                    10500,      0.1%
 - ReadSubstate                                                             ,                  1325876,     14.3%
 - RunNativeCode::AuthZone_pop                                              ,                    33696,      0.4%
@@ -55,8 +55,8 @@ Execution Cost Breakdown                                                   ,    
 - RunNativeCode::try_deposit_batch_or_abort                                ,                   121257,      1.3%
 - RunNativeCode::unlock_amount_FungibleVault                               ,                    46544,      0.5%
 - RunNativeCode::withdraw                                                  ,                    57851,      0.6%
-- RunWasmCode::BasicFlashLoan_repay_loan                                   ,                    88411,      1.0%
-- RunWasmCode::BasicFlashLoan_take_loan                                    ,                    80292,      0.9%
+- RunWasmCode::BasicFlashLoan_repay_loan                                   ,                    83700,      0.9%
+- RunWasmCode::BasicFlashLoan_take_loan                                    ,                    75263,      0.8%
 - SetCallFrameData                                                         ,                      606,      0.0%
 - SwitchStack                                                              ,                     1000,      0.0%
 - ValidateTxPayload                                                        ,                    18800,      0.2%