--- a/radix-engine-tests/tests/application/preview.rs
+++ b/radix-engine-tests/tests/application/preview.rs
@@ -79,7 +79,7 @@ fn test_transaction_preview_cost_estimate() {
             .checked_add(
                 Decimal::try_from(EXECUTION_COST_UNIT_PRICE_IN_XRD)
                     .unwrap()
-                    .checked_mul(FeeTable::new().validate_tx_payload_cost(size_diff))
+                    .checked_mul(FeeTable::latest().validate_tx_payload_cost(size_diff))
                     .unwrap()
             )
             .unwrap()
@@ -93,7 +93,7 @@ fn test_transaction_preview_cost_estimate() {
             .checked_add(
                 Decimal::try_from(EXECUTION_COST_UNIT_PRICE_IN_XRD)
                     .unwrap()
-                    .checked_mul(FeeTable::new().verify_tx_signatures_cost(2))
+                    .checked_mul(FeeTable::latest().verify_tx_signatures_cost(2))
                     .unwrap()
             )
             .unwrap(),