## Summary

This is to make sure below benchmarks are more or less balanced.

Test | Base | PR | %
-- | -- | -- | --
costing::scrypto_malloc | 452.9±0.43ms | 498.3±0.56ms | +10.02%
costing::scrypto_sbor_decode | 478.7±2.26ms | 533.2±2.06ms | +11.39%
costing::scrypto_sha256 | 354.9±0.68ms | 442.4±0.94ms | +24.65%
costing::spin_loop_v1 | 463.7±0.97ms | 433.6±4.55ms | -6.49%
costing::spin_loop_v2 | 941.7±3.20ms | 543.2±1.42ms | -42.32%